Factor out container utilities to separate module
