+++ dangerzone/isolation_provider/container.py
@@ -1,13 +1,11 @@
-import gzip
-import json
 import logging
 import os
 import platform
 import shlex
-import shutil
 import subprocess
-from typing import Dict, List, Tuple
+from typing import List
 
+from .. import container_utils, errors
 from ..document import Document
 from ..util import get_resource_path, get_subprocess_startupinfo
 from .base import IsolationProvider, terminate_process_group
@@ -26,88 +24,8 @@ else:
 log = logging.getLogger(__name__)
 
 
-class NoContainerTechException(Exception):
-    def __init__(self, container_tech: str) -> None:
-        super().__init__(f"{container_tech} is not installed")
-
-
-class NotAvailableContainerTechException(Exception):
-    def __init__(self, container_tech: str, error: str) -> None:
-        self.error = error
-        self.container_tech = container_tech
-        super().__init__(f"{container_tech} is not available")
-
-
-class ImageNotPresentException(Exception):
-    pass
-
-
-class ImageInstallationException(Exception):
-    pass
-
-
 class Container(IsolationProvider):
     # Name of the dangerzone container
-    CONTAINER_NAME = "dangerzone.rocks/dangerzone"
-
-    @staticmethod
-    def get_runtime_name() -> str:
-        if platform.system() == "Linux":
-            runtime_name = "podman"
-        else:
-            # Windows, Darwin, and unknown use docker for now, dangerzone-vm eventually
-            runtime_name = "docker"
-        return runtime_name
-
-    @staticmethod
-    def get_runtime_version() -> Tuple[int, int]:
-        """Get the major/minor parts of the Docker/Podman version.
-
-        Some of the operations we perform in this module rely on some Podman features
-        that are not available across all of our platforms. In order to have a proper
-        fallback, we need to know the Podman version. More specifically, we're fine with
-        just knowing the major and minor version, since writing/installing a full-blown
-        semver parser is an overkill.
-        """
-        # Get the Docker/Podman version, using a Go template.
-        runtime = Container.get_runtime_name()
-        if runtime == "podman":
-            query = "{{.Client.Version}}"
-        else:
-            query = "{{.Server.Version}}"
-
-        cmd = [runtime, "version", "-f", query]
-        try:
-            version = subprocess.run(
-                cmd,
-                startupinfo=get_subprocess_startupinfo(),
-                capture_output=True,
-                check=True,
-            ).stdout.decode()
-        except Exception as e:
-            msg = f"Could not get the version of the {runtime.capitalize()} tool: {e}"
-            raise RuntimeError(msg) from e
-
-        # Parse this version and return the major/minor parts, since we don't need the
-        # rest.
-        try:
-            major, minor, _ = version.split(".", 3)
-            return (int(major), int(minor))
-        except Exception as e:
-            msg = (
-                f"Could not parse the version of the {runtime.capitalize()} tool"
-                f" (found: '{version}') due to the following error: {e}"
-            )
-            raise RuntimeError(msg)
-
-    @staticmethod
-    def get_runtime() -> str:
-        container_tech = Container.get_runtime_name()
-        runtime = shutil.which(container_tech)
-        if runtime is None:
-            raise NoContainerTechException(container_tech)
-        return runtime
-
     @staticmethod
     def get_runtime_security_args() -> List[str]:
         """Security options applicable to the outer Dangerzone container.
@@ -131,7 +49,7 @@ class Container(IsolationProvider):
           - This particular argument is specified in `start_doc_to_pixels_proc()`, but
             should move here once #748 is merged.
         """
-        if Container.get_runtime_name() == "podman":
+        if container_utils.get_runtime_name() == "podman":
             security_args = ["--log-driver", "none"]
             security_args += ["--security-opt", "no-new-privileges"]
         else:
@@ -156,114 +74,6 @@ class Container(IsolationProvider):
         return security_args
 
     @staticmethod
-    def list_image_tags() -> Dict[str, str]:
-        """Get the tags of all loaded Dangerzone images.
-
-        Perform the following actions:
-        1. Get the tags of any locally available images that match Dangerzone's image
-           name.
-        2. Get the expected image tag from the image-id.txt file.
-           - If this tag is present in the local images, then we can return.
-           - Else, prune the older container images and continue.
-        3. Load the image tarball and make sure it matches the expected tag.
-        """
-        images = json.loads(
-            subprocess.check_output(
-                [
-                    Container.get_runtime(),
-                    "image",
-                    "list",
-                    "--format",
-                    "json",
-                    Container.CONTAINER_NAME,
-                ],
-                text=True,
-                startupinfo=get_subprocess_startupinfo(),
-            )
-        )
-
-        # Grab every image name and associate it with an image ID.
-        tags = {}
-        for image in images:
-            for name in image["Names"]:
-                tag = name.split(":")[1]
-                tags[tag] = image["Id"]
-
-        return tags
-
-    @staticmethod
-    def delete_image_tag(tag: str) -> None:
-        """Delete a Dangerzone image tag."""
-        name = Container.CONTAINER_NAME + ":" + tag
-        log.warning(f"Deleting old container image: {name}")
-        try:
-            subprocess.check_output(
-                [Container.get_runtime(), "rmi", "--force", name],
-                startupinfo=get_subprocess_startupinfo(),
-            )
-        except Exception as e:
-            log.warning(
-                f"Couldn't delete old container image '{name}', so leaving it there."
-                f" Original error: {e}"
-            )
-
-    @staticmethod
-    def add_image_tag(cur_tag: str, new_tag: str) -> None:
-        """Add a tag to an existing Dangerzone image."""
-        cur_image_name = Container.CONTAINER_NAME + ":" + cur_tag
-        new_image_name = Container.CONTAINER_NAME + ":" + new_tag
-        subprocess.check_output(
-            [
-                Container.get_runtime(),
-                "tag",
-                cur_image_name,
-                new_image_name,
-            ],
-            startupinfo=get_subprocess_startupinfo(),
-        )
-
-        log.info(
-            f"Successfully tagged container image '{cur_image_name}' as '{new_image_name}'"
-        )
-
-    @staticmethod
-    def get_expected_tag() -> str:
-        """Get the tag of the Dangerzone image tarball from the image-id.txt file."""
-        with open(get_resource_path("image-id.txt")) as f:
-            return f.read().strip()
-
-    @staticmethod
-    def load_image_tarball() -> None:
-        log.info("Installing Dangerzone container image...")
-        p = subprocess.Popen(
-            [Container.get_runtime(), "load"],
-            stdin=subprocess.PIPE,
-            startupinfo=get_subprocess_startupinfo(),
-        )
-
-        chunk_size = 4 << 20
-        compressed_container_path = get_resource_path("container.tar.gz")
-        with gzip.open(compressed_container_path) as f:
-            while True:
-                chunk = f.read(chunk_size)
-                if len(chunk) > 0:
-                    if p.stdin:
-                        p.stdin.write(chunk)
-                else:
-                    break
-        _, err = p.communicate()
-        if p.returncode < 0:
-            if err:
-                error = err.decode()
-            else:
-                error = "No output"
-            raise ImageInstallationException(
-                f"Could not install container image: {error}"
-            )
-
-        log.info("Successfully installed container image from")
-
-    @staticmethod
     def install() -> bool:
         """Install the container image tarball, or verify that it's already installed.
 
@@ -275,8 +85,8 @@ class Container(IsolationProvider):
            - Else, prune the older container images and continue.
         3. Load the image tarball and make sure it matches the expected tag.
         """
-        old_tags = Container.list_image_tags()
-        expected_tag = Container.get_expected_tag()
+        old_tags = container_utils.list_image_tags()
+        expected_tag = container_utils.get_expected_tag()
 
         if expected_tag not in old_tags:
             # Prune older container images.
@@ -289,14 +99,14 @@ class Container(IsolationProvider):
             return True
 
         # Load the image tarball into the container runtime.
-        Container.load_image_tarball()
+        container_utils.load_image_tarball()
 
         # Check that the container image has the expected image tag.
         # See https://github.com/freedomofpress/dangerzone/issues/988 for an example
         # where this was not the case.
-        new_tags = Container.list_image_tags()
+        new_tags = container_utils.list_image_tags()
         if expected_tag not in new_tags:
-            raise ImageNotPresentException(
+            raise errors.ImageNotPresentException(
                 f"Could not find expected tag '{expected_tag}' after loading the"
                 " container image tarball"
             )
@@ -309,8 +119,8 @@ class Container(IsolationProvider):
 
     @staticmethod
     def is_available() -> bool:
-        container_runtime = Container.get_runtime()
-        runtime_name = Container.get_runtime_name()
+        container_runtime = container_utils.get_runtime()
+        runtime_name = container_utils.get_runtime_name()
         # Can we run `docker/podman image ls` without an error
         with subprocess.Popen(
             [container_runtime, "image", "ls"],
@@ -320,7 +130,9 @@ class Container(IsolationProvider):
         ) as p:
             _, stderr = p.communicate()
             if p.returncode != 0:
-                raise NotAvailableContainerTechException(runtime_name, stderr.decode())
+                raise errors.NotAvailableContainerTechException(
+                    runtime_name, stderr.decode()
+                )
             return True
 
     def doc_to_pixels_container_name(self, document: Document) -> str:
@@ -355,7 +167,7 @@ class Container(IsolationProvider):
         name: str,
         extra_args: List[str] = [],
     ) -> subprocess.Popen:
-        container_runtime = self.get_runtime()
+        container_runtime = container_utils.get_runtime()
         security_args = self.get_runtime_security_args()
         enable_stdin = ["-i"]
         set_name = ["--name", name]
@@ -385,7 +197,7 @@ class Container(IsolationProvider):
         connected to the Docker daemon, and killing it will just close the associated
         standard streams.
         """
-        container_runtime = self.get_runtime()
+        container_runtime = container_utils.get_runtime()
         cmd = [container_runtime, "kill", name]
         try:
             # We do not check the exit code of the process here, since the container may
@@ -421,8 +233,8 @@ class Container(IsolationProvider):
         # NOTE: Using `--userns nomap` is available only on Podman >= 4.1.0.
         # XXX: Move this under `get_runtime_security_args()` once #748 is merged.
         extra_args = []
-        if Container.get_runtime_name() == "podman":
-            if Container.get_runtime_version() >= (4, 1):
+        if container_utils.get_runtime_name() == "podman":
+            if container_utils.get_runtime_version() >= (4, 1):
                 extra_args += ["--userns", "nomap"]
 
         name = self.doc_to_pixels_container_name(document)
@@ -449,7 +261,7 @@ class Container(IsolationProvider):
         # after a podman kill / docker kill invocation, this will likely be the case,
         # else the container runtime (Docker/Podman) has experienced a problem, and we
         # should report it.
-        container_runtime = self.get_runtime()
+        container_runtime = container_utils.get_runtime()
         name = self.doc_to_pixels_container_name(document)
         all_containers = subprocess.run(
             [container_runtime, "ps", "-a"],
@@ -471,11 +283,11 @@ class Container(IsolationProvider):
             if cpu_count is not None:
                 n_cpu = cpu_count
 
-        elif self.get_runtime_name() == "docker":
+        elif container_utils.get_runtime_name() == "docker":
             # For Windows and MacOS containers run in VM
             # So we obtain the CPU count for the VM
             n_cpu_str = subprocess.check_output(
-                [self.get_runtime(), "info", "--format", "{{.NCPU}}"],
+                [container_utils.get_runtime(), "info", "--format", "{{.NCPU}}"],
                 text=True,
                 startupinfo=get_subprocess_startupinfo(),
             )
